pyproject.toml
[project]
name = "goldenverba"
version = "2.1.3"
description = "Welcome to Verba: The Golden RAGtriever, an open-source initiative designed to offer a streamlined, user-friendly interface for Retrieval-Augmented Generation (RAG) applications. In just a few easy steps, dive into your data and make meaningful interactions!"
readme = "README.md"
authors = [
    {name = "Weaviate", email = "edward@weaviate.io"}
]
license = {text = "BSD License"}
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["RAG", "retrieval", "AI", "vector", "search", "weaviate"]
requires-python = ">=3.10, <3.15"
dependencies = [
    "weaviate-client>=4.9.6",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.5",
    "wasabi>=1.1.2",
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.0",
    "gunicorn>=22.0.0",
    "click>=8.1.7",
    "xlrd>=2.0.2",
    "tiktoken>=0.9.0",
    "requests>=2.32.3",
    "pypdf>=4.3.1",
    "python-docx>=1.1.2",
    "scikit-learn>=1.5.2",
    "langchain-text-splitters>=0.2.2",
    "spacy>=3.7.5",
    "aiohttp>=3.10.2",
    "httpx>=0.27.0",
    "markdownify>=0.13.1",
    "aiofiles>=24.1.0",
    "assemblyai>=0.33.0",
    "beautifulsoup4>=4.12.3",
    "langdetect>=1.0.9",
    # Ensure compiled extensions (thinc/spaCy) use a compatible NumPy ABI
    "numpy>=1.26,<2.0",
    # Ensure compatible protobuf runtime across deps (Weaviate, Vertex AI)
    # google-cloud-aiplatform/vertexai currently require <5; allow 4.25.3+
    # while remaining forward-compatible with future <6 releases.
    "protobuf>=4.25.3,<6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "aioresponses>=0.7.4",
    "wheel",
    "twine",
    "black>=23.7.0",
    "setuptools",
    "ruff>=0.4.0",
]
google = [
    "vertexai==1.46.0",
]
huggingface = [
    "sentence-transformers==3.0.1",
]

[project.scripts]
verba = "goldenverba.server.cli:cli"

[project.urls]
Homepage = "https://github.com/weaviate/Verba"
Repository = "https://github.com/weaviate/Verba"
Documentation = "https://github.com/weaviate/Verba"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "I",      # isort
]
ignore = [
    "COM812", # trailing comma conflicts with formatter
    "ISC001", # single line implicit string concatenation
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]  # Allow assert in tests
"**/test_*.py" = ["S101"]   # Allow assert in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["goldenverba"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | venv
  )/
)
'''

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "goldenverba/tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"


# Ty: extremely fast Python type checker (preview)
# See: https://docs.astral.sh/ty/
[tool.ty]

[tool.hatch.build]
include = [
  "goldenverba/**",
  "README.md",
  "LICENSE",
]
